fix(griffin): cast char to String or Symbol could produce unprintable strings

+++ core/src/main/java/io/questdb/cairo/CairoConfiguration.java
@@ -100,9 +100,9 @@ public interface CairoConfiguration {
 
     int getSqlModelPoolCapacity();
 
-    int getSqlSortKeyPageSize();
+    long getSqlSortKeyPageSize();
 
-    int getSqlSortLightValuePageSize();
+    long getSqlSortLightValuePageSize();
 
     int getSqlHashJoinValuePageSize();
